requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
kafka-python==2.0.2
praw==7.7.0
prawcore==2.3.0
py4j==0.10.9.7
pyspark==3.4.0
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
update-checker==0.18.0
urllib3==2.0.2
websocket-client==1.5.1